+++ odoo/local-src/specific_purchase/models/purchase.py
@@ -61,15 +61,3 @@ class PurchaseOrderLine(models.Model):
     req_sn_supplier = fields.Char(
         string='Requested SN to supplier',
     )
-
-
-class StocProduction(models.Model):
-    _inherit = 'stock.production.lot'
-
-    @api.model
-    def name_search(self, name='', args=None, operator='ilike', limit=100):
-        return super(StocProduction, self).name_search(
-            name=name,
-            args=args,
-            operator=operator,
-            limit=limit)